+++ VesselPlanner/ToolbarRegistration.cs
@@ -1,14 +1,50 @@
-﻿using VesselPlanner.UI;
-using KSP.UI.Screens;
-using ToolbarControl_NS;
+﻿using ToolbarControl_NS;
 using UnityEngine;
+using VesselPlanner.UI;
 
 namespace VesselPlanner
 {
     [KSPAddon(KSPAddon.Startup.MainMenu, true)]
     public sealed class ToolbarRegistration : MonoBehaviour
     {
+
+        internal static Texture2D texDarker, texLighter;
+        internal static GUIStyle winDarker, winLighter;
         bool initialized = false;
+
+        void InitWinTextures()
+        {
+            float shade = 0.12f;
+            Color color = new Color(shade, shade, shade, 1f);
+            winDarker = new GUIStyle(HighLogic.Skin.window);
+            texDarker = winDarker.normal.background;
+            var pixels = texDarker.GetPixels32();
+            for (int i = 0; i < pixels.Length; ++i)
+                pixels[i].a = 255;
+            texDarker.SetPixels32(pixels); texDarker.Apply();
+            winDarker.active.background =
+            winDarker.focused.background =
+            winDarker.normal.background = texDarker;
+
+            // --------------------------
+
+            shade = 0.36f;
+            color = new Color(shade, shade, shade, 1f);
+            winLighter = new GUIStyle(HighLogic.Skin.window);
+            texLighter = winLighter.normal.background;
+            pixels = texLighter.GetPixels32();
+            for (int i = 0; i < pixels.Length; ++i)
+                pixels[i].a = 255;
+            texLighter.SetPixels32(pixels); texLighter.Apply();
+            winLighter.active.background =
+            winLighter.focused.background =
+            winLighter.normal.background = texLighter;
+
+            winLighter.active.background =
+            winLighter.focused.background =
+            winLighter.normal.background = texLighter;
+        }
+
         private void Start()
         {
             ToolbarControl.RegisterMod(VesselPlannerEditorAddon.MODID, VesselPlannerEditorAddon.MODNAME);
@@ -27,6 +63,8 @@ namespace VesselPlanner
                 PlannerWindow.EnsureStyles();
                 FlightGraphWindow.InitStyles();
                 initialized = true;
+
+                InitWinTextures();
             }
         }
     }